+++ R/reformat.R
@@ -86,4 +86,9 @@
     }
   }
   tinkr::to_md(md_doc, path = file)
+
+  # 4 spaces for nested lists (can't find a way to fix this in XML)
+  doc <- readLines(file, warn = FALSE)
+  new_doc <- gsub("^  (\\*|-)", "    \\1", doc)
+  cat(paste(new_doc, collapse = "\n"), file = file)
 }