+++ docs/NEWS.md
@@ -1,3 +1,8 @@
+## altdoc (development version)
+
+- If necessary, two spaces are automatically added in nested lists n the `NEWS`
+  (or `Changelog`) file.
+
 ## altdoc 0.2.1
 
 - Fix test failures on CRAN due to the new version of `usethis`
@@ -35,4 +40,3 @@
 ## altdoc 0.1.0
 
 - First version.
-